requirements.txt
bcrypt
cryptography
argon2-cffi